Refactor integration of asset FT with the DEX to accept all `DEXActions` by asset FT required to be applied.
