+++ x/asset/ft/keeper/keeper.go
@@ -18,8 +18,6 @@ import (
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-	"github.com/pkg/errors"
-	"github.com/samber/lo"
 
 	"github.com/CoreumFoundation/coreum/v5/x/asset/ft/types"
 	"github.com/CoreumFoundation/coreum/v5/x/wasm"
@@ -704,287 +702,6 @@ func (k Keeper) SetWhitelistedBalances(ctx sdk.Context, addr sdk.AccAddress, coi
 	}
 }
 
-// DEXIncreaseLimits increases the DEX limits.
-func (k Keeper) DEXIncreaseLimits(
-	ctx sdk.Context, addr sdk.AccAddress, lockedCoin, expectedToReceiveCoin sdk.Coin,
-) error {
-	if err := k.dexChecksForDenoms(ctx, addr, lockedCoin.Denom, expectedToReceiveCoin.Denom); err != nil {
-		return err
-	}
-
-	if err := k.DEXLock(ctx, addr, lockedCoin); err != nil {
-		return err
-	}
-
-	return k.DEXIncreaseExpectedToReceive(ctx, addr, expectedToReceiveCoin)
-}
-
-// DEXDecreaseLimits decreases the DEX limits.
-func (k Keeper) DEXDecreaseLimits(
-	ctx sdk.Context,
-	addr sdk.AccAddress,
-	lockedCoin, expectedToReceiveCoin sdk.Coin,
-) error {
-	if err := k.DEXUnlock(ctx, addr, lockedCoin); err != nil {
-		return err
-	}
-
-	return k.DEXDecreaseExpectedToReceive(ctx, addr, expectedToReceiveCoin)
-}
-
-// DEXDecreaseLimitsAndSend decreases the DEX limits and sends the coin.
-func (k Keeper) DEXDecreaseLimitsAndSend(
-	ctx sdk.Context, fromAddr, toAddr sdk.AccAddress, unlockAndSendCoin, decreaseExpectedToReceiveCoin sdk.Coin,
-) error {
-	if err := k.DEXUnlock(ctx, fromAddr, unlockAndSendCoin); err != nil {
-		return err
-	}
-	if err := k.DEXDecreaseExpectedToReceive(ctx, fromAddr, decreaseExpectedToReceiveCoin); err != nil {
-		return err
-	}
-	// send using the native bank
-	if err := k.bankKeeper.SendCoins(ctx, fromAddr, toAddr, sdk.NewCoins(unlockAndSendCoin)); err != nil {
-		return sdkerrors.Wrap(err, "failed to send DEX coins")
-	}
-
-	return nil
-}
-
-// DEXCheckLimitsAndSend checks DEX limits and sends the coin.
-func (k Keeper) DEXCheckLimitsAndSend(
-	ctx sdk.Context, fromAddr, toAddr sdk.AccAddress, sendCoin, checkExpectedToReceiveCoin sdk.Coin,
-) error {
-	if err := k.dexChecksForDenoms(ctx, fromAddr, sendCoin.Denom, checkExpectedToReceiveCoin.Denom); err != nil {
-		return err
-	}
-
-	if err := k.dexLockingChecks(ctx, fromAddr, sendCoin); err != nil {
-		return err
-	}
-
-	// check the whitelisted balance only if we record the expected to receive amount
-	shouldCheckWhitelistedBalance, err := k.shouldRecordExpectedToReceiveBalance(
-		ctx, fromAddr, checkExpectedToReceiveCoin.Denom,
-	)
-	if err != nil {
-		return err
-	}
-	if shouldCheckWhitelistedBalance {
-		if err := k.validateWhitelistedBalance(ctx, fromAddr, checkExpectedToReceiveCoin); err != nil {
-			return err
-		}
-	}
-
-	if err := k.bankKeeper.SendCoins(ctx, fromAddr, toAddr, sdk.NewCoins(sendCoin)); err != nil {
-		return sdkerrors.Wrap(err, "failed to DEX send coins")
-	}
-
-	return nil
-}
-
-// DEXIncreaseExpectedToReceive increases the expected to receive amount for the specified account.
-func (k Keeper) DEXIncreaseExpectedToReceive(ctx sdk.Context, addr sdk.AccAddress, coin sdk.Coin) error {
-	if !coin.IsPositive() {
-		return sdkerrors.Wrap(
-			cosmoserrors.ErrInvalidCoins, "amount to increase DEX expected to receive must be positive",
-		)
-	}
-
-	shouldRecord, err := k.shouldRecordExpectedToReceiveBalance(ctx, addr, coin.Denom)
-	if err != nil {
-		return err
-	}
-	if !shouldRecord {
-		return nil
-	}
-
-	if err := k.validateWhitelistedBalance(ctx, addr, coin); err != nil {
-		return err
-	}
-
-	dexExpectedToReceiveStore := k.dexExpectedToReceiveAccountBalanceStore(ctx, addr)
-	prevExpectedToReceiveBalance, newExpectedToReceiveBalance := dexExpectedToReceiveStore.AddBalance(coin)
-
-	if err := ctx.EventManager().EmitTypedEvent(&types.EventDEXExpectedToReceiveAmountChanged{
-		Account:        addr.String(),
-		Denom:          coin.Denom,
-		PreviousAmount: prevExpectedToReceiveBalance.Amount,
-		CurrentAmount:  newExpectedToReceiveBalance.Amount,
-	}); err != nil {
-		return sdkerrors.Wrapf(
-			types.ErrInvalidState, "failed to emit EventDEXExpectedToReceiveAmountChanged event: %s", err,
-		)
-	}
-
-	return nil
-}
-
-// DEXDecreaseExpectedToReceive decreases the expected to receive amount for the specified account.
-func (k Keeper) DEXDecreaseExpectedToReceive(ctx sdk.Context, addr sdk.AccAddress, coin sdk.Coin) error {
-	if !coin.IsPositive() {
-		return sdkerrors.Wrap(
-			cosmoserrors.ErrInvalidCoins, "amount to decrease DEX expected to receive must be positive",
-		)
-	}
-
-	shouldRecord, err := k.shouldRecordExpectedToReceiveBalance(ctx, addr, coin.Denom)
-	if err != nil {
-		return err
-	}
-	if !shouldRecord {
-		return nil
-	}
-
-	dexExpectedToReceiveStore := k.dexExpectedToReceiveAccountBalanceStore(ctx, addr)
-	prevExpectedToReceiveBalance, newExpectedToReceiveBalance, err := dexExpectedToReceiveStore.SubBalance(coin)
-	if err != nil {
-		return sdkerrors.Wrap(err, "failed to cancel DEX whitelisted")
-	}
-
-	if err := ctx.EventManager().EmitTypedEvent(&types.EventDEXExpectedToReceiveAmountChanged{
-		Account:        addr.String(),
-		Denom:          coin.Denom,
-		PreviousAmount: prevExpectedToReceiveBalance.Amount,
-		CurrentAmount:  newExpectedToReceiveBalance.Amount,
-	}); err != nil {
-		return sdkerrors.Wrapf(
-			types.ErrInvalidState, "failed to emit EventDEXExpectedToReceiveAmountChanged event: %s", err,
-		)
-	}
-
-	return nil
-}
-
-// GetDEXExpectedToReceivedBalance returns the DEX expected to receive balance.
-func (k Keeper) GetDEXExpectedToReceivedBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) sdk.Coin {
-	return k.dexExpectedToReceiveAccountBalanceStore(ctx, addr).Balance(denom)
-}
-
-// GetDEXExpectedToReceiveBalances returns the DEX expected to receive balances of an account.
-func (k Keeper) GetDEXExpectedToReceiveBalances(
-	ctx sdk.Context,
-	addr sdk.AccAddress,
-	pagination *query.PageRequest,
-) (sdk.Coins, *query.PageResponse, error) {
-	return k.dexExpectedToReceiveAccountBalanceStore(ctx, addr).Balances(pagination)
-}
-
-// GetAccountsDEXExpectedToReceiveBalances returns the DEX expected to receive balance on all the account.
-func (k Keeper) GetAccountsDEXExpectedToReceiveBalances(
-	ctx sdk.Context,
-	pagination *query.PageRequest,
-) ([]types.Balance, *query.PageResponse, error) {
-	return collectBalances(k.cdc, k.dexExpectedToReceiveBalancesStore(ctx), pagination)
-}
-
-// SetDEXExpectedToReceiveBalances sets the DEX expected to receive balances of a specified account.
-// Pay attention that the sdk.NewCoins() sanitizes/removes the empty coins, hence if you
-// need set zero amount use the slice []sdk.Coins.
-func (k Keeper) SetDEXExpectedToReceiveBalances(ctx sdk.Context, addr sdk.AccAddress, coins sdk.Coins) {
-	dexExpectedToReceiveStore := k.dexExpectedToReceiveAccountBalanceStore(ctx, addr)
-	for _, coin := range coins {
-		dexExpectedToReceiveStore.SetBalance(coin)
-	}
-}
-
-// DEXLock locks specified token for the specified account.
-func (k Keeper) DEXLock(ctx sdk.Context, addr sdk.AccAddress, coin sdk.Coin) error {
-	if err := k.dexLockingChecks(ctx, addr, coin); err != nil {
-		return err
-	}
-
-	dexLockedStore := k.dexLockedAccountBalanceStore(ctx, addr)
-	prevLockedBalance, newLockedBalance := dexLockedStore.AddBalance(coin)
-
-	if err := ctx.EventManager().EmitTypedEvent(&types.EventDEXLockedAmountChanged{
-		Account:        addr.String(),
-		Denom:          coin.Denom,
-		PreviousAmount: prevLockedBalance.Amount,
-		CurrentAmount:  newLockedBalance.Amount,
-	}); err != nil {
-		return sdkerrors.Wrapf(types.ErrInvalidState, "failed to emit EventDEXLockedAmountChanged event: %s", err)
-	}
-
-	return nil
-}
-
-// DEXUnlock unlocks specified tokens from the specified account.
-func (k Keeper) DEXUnlock(ctx sdk.Context, addr sdk.AccAddress, coin sdk.Coin) error {
-	if !coin.IsPositive() {
-		return sdkerrors.Wrap(cosmoserrors.ErrInvalidCoins, "amount to unlock DEX tokens must be positive")
-	}
-
-	dexLockedStore := k.dexLockedAccountBalanceStore(ctx, addr)
-	prevLockedBalance, newLockedBalance, err := dexLockedStore.SubBalance(coin)
-	if err != nil {
-		return sdkerrors.Wrap(err, "failed to unlock DEX")
-	}
-
-	if err := ctx.EventManager().EmitTypedEvent(&types.EventDEXLockedAmountChanged{
-		Account:        addr.String(),
-		Denom:          coin.Denom,
-		PreviousAmount: prevLockedBalance.Amount,
-		CurrentAmount:  newLockedBalance.Amount,
-	}); err != nil {
-		return sdkerrors.Wrapf(types.ErrInvalidState, "failed to emit EventDEXLockedAmountChanged event: %s", err)
-	}
-
-	return nil
-}
-
-// GetDEXLockedBalance returns the DEX locked balance.
-func (k Keeper) GetDEXLockedBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) sdk.Coin {
-	return k.dexLockedAccountBalanceStore(ctx, addr).Balance(denom)
-}
-
-// GetDEXLockedBalances returns the DEX locked balances of an account.
-func (k Keeper) GetDEXLockedBalances(
-	ctx sdk.Context,
-	addr sdk.AccAddress,
-	pagination *query.PageRequest,
-) (sdk.Coins, *query.PageResponse, error) {
-	return k.dexLockedAccountBalanceStore(ctx, addr).Balances(pagination)
-}
-
-// GetAccountsDEXLockedBalances returns the DEX locked balance on all the account.
-func (k Keeper) GetAccountsDEXLockedBalances(
-	ctx sdk.Context,
-	pagination *query.PageRequest,
-) ([]types.Balance, *query.PageResponse, error) {
-	return collectBalances(k.cdc, k.dexLockedBalancesStore(ctx), pagination)
-}
-
-// SetDEXLockedBalances sets the DEX locked balances of a specified account.
-// Pay attention that the sdk.NewCoins() sanitizes/removes the empty coins, hence if you
-// need set zero amount use the slice []sdk.Coins.
-func (k Keeper) SetDEXLockedBalances(ctx sdk.Context, addr sdk.AccAddress, coins sdk.Coins) {
-	dexLockedStore := k.dexLockedAccountBalanceStore(ctx, addr)
-	for _, coin := range coins {
-		dexLockedStore.SetBalance(coin)
-	}
-}
-
-// ValidateDEXCancelOrdersByDenomIsAllowed validates whether the cancellation of  orders by denom is allowed.
-func (k Keeper) ValidateDEXCancelOrdersByDenomIsAllowed(ctx sdk.Context, addr sdk.AccAddress, denom string) error {
-	def, err := k.GetDefinition(ctx, denom)
-	if err != nil {
-		return err
-	}
-
-	if !def.HasAdminPrivileges(addr) {
-		return sdkerrors.Wrapf(cosmoserrors.ErrUnauthorized, "only admin is able to cancel orders by denom %s", denom)
-	}
-	if !def.IsFeatureEnabled(types.Feature_dex_order_cancellation) {
-		return sdkerrors.Wrapf(
-			cosmoserrors.ErrUnauthorized,
-			"order cancellation is not allowed by denom %s, feature %s is disabled",
-			denom, types.Feature_dex_order_cancellation,
-		)
-	}
-
-	return nil
-}
-
 // GetSpendableBalance returns balance allowed to be spent.
 func (k Keeper) GetSpendableBalance(
 	ctx sdk.Context,
@@ -1002,7 +719,6 @@ func (k Keeper) GetSpendableBalance(
 	if notLockedAmt.IsNegative() {
 		return sdk.NewCoin(denom, sdkmath.ZeroInt())
 	}
-
 	notFrozenAmt := balance.Amount.Sub(k.GetFrozenBalance(ctx, addr, denom).Amount)
 	if notFrozenAmt.IsNegative() {
 		return sdk.NewCoin(denom, sdkmath.ZeroInt())
@@ -1076,130 +792,6 @@ func (k Keeper) ClearAdmin(ctx sdk.Context, sender sdk.AccAddress, denom string)
 		PreviousAdmin: previousAdmin,
 	}); err != nil {
 		return sdkerrors.Wrapf(types.ErrInvalidState, "failed to emit EventAdminCleared event: %s", err)
-	}
-
-	return nil
-}
-
-// SetDEXSettings sets the DEX settings of a specified denom.
-func (k Keeper) SetDEXSettings(ctx sdk.Context, denom string, settings types.DEXSettings) {
-	ctx.KVStore(k.storeKey).Set(types.CreateDEXSettingsKey(denom), k.cdc.MustMarshal(&settings))
-}
-
-// GetDEXSettings gets the DEX settings of a specified denom.
-func (k Keeper) GetDEXSettings(ctx sdk.Context, denom string) (types.DEXSettings, error) {
-	bz := ctx.KVStore(k.storeKey).Get(types.CreateDEXSettingsKey(denom))
-	if bz == nil {
-		return types.DEXSettings{}, sdkerrors.Wrapf(types.ErrDEXSettingsNotFound, "denom: %s", denom)
-	}
-	var settings types.DEXSettings
-	k.cdc.MustUnmarshal(bz, &settings)
-
-	return settings, nil
-}
-
-// GetDEXSettingsWithDenoms returns all DEX settings with the corresponding denoms.
-func (k Keeper) GetDEXSettingsWithDenoms(
-	ctx sdk.Context,
-	pagination *query.PageRequest,
-) ([]types.DEXSettingsWithDenom, *query.PageResponse, error) {
-	dexSettings := make([]types.DEXSettingsWithDenom, 0)
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.DEXSettingsKeyPrefix)
-	pageRes, err := query.Paginate(store, pagination, func(key, value []byte) error {
-		denom, err := types.DecodeDenomFromKey(key)
-		if err != nil {
-			return err
-		}
-		var settings types.DEXSettings
-		k.cdc.MustUnmarshal(value, &settings)
-
-		dexSettings = append(dexSettings, types.DEXSettingsWithDenom{
-			Denom:       denom,
-			DEXSettings: settings,
-		})
-
-		return nil
-	})
-
-	return dexSettings, pageRes, err
-}
-
-// UpdateDEXUnifiedRefAmount updates the DEX unified ref amount .
-func (k Keeper) UpdateDEXUnifiedRefAmount(
-	ctx sdk.Context,
-	sender sdk.AccAddress,
-	denom string,
-	unifiedRefAmount sdkmath.LegacyDec,
-) error {
-	return k.updateDEXSettings(ctx, sender, denom, types.DEXSettings{UnifiedRefAmount: &unifiedRefAmount})
-}
-
-// UpdateDEXWhitelistedDenoms updates the DEX whitelisted of a specified denoms.
-func (k Keeper) UpdateDEXWhitelistedDenoms(
-	ctx sdk.Context,
-	sender sdk.AccAddress,
-	denom string,
-	whitelistedDenoms []string,
-) error {
-	if whitelistedDenoms == nil {
-		// check to prevent mistakes using the `updateDEXSettings` method, set to empty slice if the input is nil
-		whitelistedDenoms = make([]string, 0)
-	}
-	return k.updateDEXSettings(ctx, sender, denom, types.DEXSettings{WhitelistedDenoms: whitelistedDenoms})
-}
-
-func (k Keeper) updateDEXSettings(
-	ctx sdk.Context,
-	sender sdk.AccAddress,
-	denom string,
-	settings types.DEXSettings,
-) error {
-	prevSettings, err := k.getDEXSettingsOrNil(ctx, denom)
-	if err != nil {
-		return err
-	}
-	if prevSettings == nil {
-		prevSettings = &types.DEXSettings{}
-	}
-
-	newSettings := *prevSettings
-	// update not nil settings
-	if settings.WhitelistedDenoms != nil {
-		newSettings.WhitelistedDenoms = settings.WhitelistedDenoms
-	}
-	if settings.UnifiedRefAmount != nil {
-		newSettings.UnifiedRefAmount = settings.UnifiedRefAmount
-	}
-
-	if err := types.ValidateDEXSettings(settings); err != nil {
-		return err
-	}
-
-	def, err := k.getDefinitionOrNil(ctx, denom)
-	if err != nil {
-		return err
-	}
-	// the gov can update any DEX setting even if the features are disabled
-	if k.authority != sender.String() { //nolint:nestif // the ifs are for the error checks mostly
-		if def != nil {
-			if !def.IsAdmin(sender) {
-				return sdkerrors.Wrap(cosmoserrors.ErrUnauthorized, "only admin and gov can update DEX settings")
-			}
-			if err := types.ValidateDEXSettingsAccess(newSettings, *def); err != nil {
-				return err
-			}
-		} else {
-			return sdkerrors.Wrap(cosmoserrors.ErrUnauthorized, "only admin or gov can update DEX settings")
-		}
-	}
-
-	k.SetDEXSettings(ctx, denom, newSettings)
-
-	if err := ctx.EventManager().EmitTypedEvent(&types.EventDEXSettingsChanged{
-		PreviousSettings: prevSettings,
-		NewSettings:      newSettings,
-	}); err != nil {
-		return sdkerrors.Wrapf(types.ErrInvalidState, "failed to emit EventDEXSettingsChanged event: %s", err)
 	}
 
 	return nil
@@ -1565,32 +1157,6 @@ func (k Keeper) whitelistedAccountBalanceStore(ctx sdk.Context, addr sdk.AccAddr
 	return newBalanceStore(k.cdc, ctx.KVStore(k.storeKey), types.CreateWhitelistedBalancesKey(addr))
 }
 
-// dexExpectedToReceiveBalancesStore get the store for the DEX expected to receive balances of all accounts.
-func (k Keeper) dexExpectedToReceiveBalancesStore(ctx sdk.Context) prefix.Store {
-	return prefix.NewStore(ctx.KVStore(k.storeKey), types.DEXExpectedToReceiveBalancesKeyPrefix)
-}
-
-// dexExpectedToReceiveAccountBalanceStore gets the store for the DEX expected to receive balances of an account.
-func (k Keeper) dexExpectedToReceiveAccountBalanceStore(ctx sdk.Context, addr sdk.AccAddress) balanceStore {
-	return newBalanceStore(k.cdc, ctx.KVStore(k.storeKey), types.CreateDEXExpectedToReceiveBalancesKey(addr))
-}
-
-func (k Keeper) shouldRecordExpectedToReceiveBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) (bool, error) {
-	def, err := k.getDefinitionOrNil(ctx, denom)
-	if err != nil {
-		return false, err
-	}
-	// increase for FT with the whitelisting enabled
-	if def == nil ||
-		!def.IsFeatureEnabled(types.Feature_whitelisting) ||
-		// it's prohibited to whitelist the admin
-		def.IsAdmin(addr) {
-		return false, nil
-	}
-
-	return true, nil
-}
-
 func (k Keeper) validateWhitelistedBalance(ctx sdk.Context, addr sdk.AccAddress, coin sdk.Coin) error {
 	balance := k.bankKeeper.GetBalance(ctx, addr, coin.Denom)
 	whitelistedBalance := k.GetWhitelistedBalance(ctx, addr, coin.Denom)
@@ -1607,163 +1173,6 @@ func (k Keeper) validateWhitelistedBalance(ctx sdk.Context, addr sdk.AccAddress,
 	}
 
 	return nil
-}
-
-// dexLockedBalancesStore get the store for the DEX locked balances of all accounts.
-func (k Keeper) dexLockedBalancesStore(ctx sdk.Context) prefix.Store {
-	return prefix.NewStore(ctx.KVStore(k.storeKey), types.DEXLockedBalancesKeyPrefix)
-}
-
-// dexLockedAccountBalanceStore gets the store for the DEX locked balances of an account.
-func (k Keeper) dexLockedAccountBalanceStore(ctx sdk.Context, addr sdk.AccAddress) balanceStore {
-	return newBalanceStore(k.cdc, ctx.KVStore(k.storeKey), types.CreateDEXLockedBalancesKey(addr))
-}
-
-func (k Keeper) dexLockingChecks(ctx sdk.Context, addr sdk.AccAddress, coin sdk.Coin) error {
-	if !coin.IsPositive() {
-		return sdkerrors.Wrap(cosmoserrors.ErrInvalidCoins, "amount to lock DEX tokens must be positive")
-	}
-
-	balance := k.bankKeeper.GetBalance(ctx, addr, coin.Denom)
-	if err := k.validateCoinIsNotLockedByDEXAndBank(ctx, addr, balance, coin); err != nil {
-		return sdkerrors.Wrapf(types.ErrDEXLockFailed, "%s", err)
-	}
-
-	frozenAmt := k.GetFrozenBalance(ctx, addr, coin.Denom).Amount
-	notFrozenTotalAmt := balance.Amount.Sub(frozenAmt)
-	if notFrozenTotalAmt.LT(coin.Amount) {
-		return sdkerrors.Wrapf(
-			types.ErrDEXLockFailed,
-			"failed to DEX lock %s available %s%s",
-			coin.String(),
-			notFrozenTotalAmt,
-			coin.Denom,
-		)
-	}
-
-	return nil
-}
-
-func (k Keeper) dexChecksForDenoms(
-	ctx sdk.Context, acc sdk.AccAddress, spendDenom, receiveDenom string,
-) error {
-	denoms := []string{spendDenom, receiveDenom}
-	for _, denom := range denoms {
-		def, err := k.getDefinitionOrNil(ctx, denom)
-		if err != nil {
-			return err
-		}
-
-		if err := k.dexChecksForDefinition(ctx, acc, def); err != nil {
-			return err
-		}
-
-		// settings specific validation
-		settings, err := k.getDEXSettingsOrNil(ctx, denom)
-		if err != nil {
-			return err
-		}
-
-		if settings != nil {
-			// validate whitelisted denoms
-			for _, tradeDenom := range denoms {
-				if denom == tradeDenom || len(settings.WhitelistedDenoms) == 0 {
-					continue
-				}
-				if !lo.Contains(settings.WhitelistedDenoms, tradeDenom) {
-					return sdkerrors.Wrapf(
-						cosmoserrors.ErrUnauthorized,
-						"locking coins for DEX is prohibited, denom %s not whitelisted for %s",
-						tradeDenom, denom,
-					)
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-func (k Keeper) dexChecksForDefinition(ctx sdk.Context, acc sdk.AccAddress, def *types.Definition) error {
-	if def == nil {
-		return nil
-	}
-
-	if def.ExtensionCWAddress != "" {
-		return sdkerrors.Wrapf(
-			types.ErrInvalidInput,
-			"usage of %s is not supported for DEX, the token has extensions",
-			def.Denom,
-		)
-	}
-
-	if def.IsFeatureEnabled(types.Feature_dex_block) {
-		return sdkerrors.Wrapf(
-			cosmoserrors.ErrUnauthorized,
-			"usage of %s is not supported for DEX, the token has %s feature enabled",
-			def.Denom, types.Feature_dex_block.String(),
-		)
-	}
-
-	// don't allow the smart contract to use the denom with Feature_block_smart_contracts if not admin
-	if def.IsFeatureEnabled(types.Feature_block_smart_contracts) &&
-		!def.HasAdminPrivileges(acc) &&
-		cwasmtypes.IsTriggeredBySmartContract(ctx) {
-		return sdkerrors.Wrapf(
-			cosmoserrors.ErrUnauthorized,
-			"usage of %s is not supported for DEX in smart contract, the token has %s feature enabled",
-			def.Denom, types.Feature_block_smart_contracts.String(),
-		)
-	}
-
-	if def.IsFeatureEnabled(types.Feature_freezing) {
-		if k.isGloballyFrozen(ctx, def.Denom) &&
-			// sill allow the admin to do the trade, to follow same logic as we have in the sending
-			!def.HasAdminPrivileges(acc) {
-			return sdkerrors.Wrapf(
-				cosmoserrors.ErrUnauthorized,
-				"usage of %s for DEX is blocked because the token is globally frozen",
-				def.Denom,
-			)
-		}
-	}
-
-	return nil
-}
-
-func (k Keeper) validateCoinIsNotLockedByDEXAndBank(
-	ctx sdk.Context,
-	addr sdk.AccAddress,
-	balance, coin sdk.Coin,
-) error {
-	dexLockedAmt := k.GetDEXLockedBalance(ctx, addr, coin.Denom).Amount
-	availableAmt := balance.Amount.Sub(dexLockedAmt)
-	if availableAmt.LT(coin.Amount) {
-		return sdkerrors.Wrapf(cosmoserrors.ErrInsufficientFunds, "%s is not available, available %s%s",
-			coin.String(), availableAmt.String(), coin.Denom)
-	}
-
-	bankLockedAmt := k.bankKeeper.LockedCoins(ctx, addr).AmountOf(coin.Denom)
-	// validate that we don't use the coins locked by bank
-	availableAmt = availableAmt.Sub(bankLockedAmt)
-	if availableAmt.LT(coin.Amount) {
-		return sdkerrors.Wrapf(cosmoserrors.ErrInsufficientFunds, "%s is not available, available %s%s",
-			coin.String(), availableAmt.String(), coin.Denom)
-	}
-
-	return nil
-}
-
-func (k Keeper) getDEXSettingsOrNil(ctx sdk.Context, denom string) (*types.DEXSettings, error) {
-	dexSettings, err := k.GetDEXSettings(ctx, denom)
-	if err != nil {
-		if errors.Is(err, types.ErrDEXSettingsNotFound) {
-			return nil, nil //nolint:nilnil //returns nil if data not found
-		}
-		return nil, err
-	}
-
-	return &dexSettings, nil
 }
 
 // logger returns the Keeper logger.
